cleanup interactions
