+++ libs/deck-layer-composer/src/interactions/deck-layers-interaction.hooks.ts
@@ -4,14 +4,7 @@ import { DeckLayerInteractionFeature } from '../types'
 import { ExtendedFeature, InteractionEvent, InteractionEventCallback } from './types'
 import { filterUniqueFeatureInteraction, getExtendedFeatures } from './interaction-features.utils'
 
-// Atom used to have all the layer instances loading state available
-export type DeckLayerInteraction = {
-  latitude: number
-  longitude: number
-  features: DeckLayerInteractionFeature[]
-}
-
-export const deckHoverInteractionAtom = atom<DeckLayerInteraction>({} as DeckLayerInteraction)
+export const deckHoverInteractionAtom = atom<InteractionEvent>({} as InteractionEvent)
 
 export const useMapHoverInteraction = () => {
   return useAtomValue(deckHoverInteractionAtom)
@@ -24,7 +17,7 @@ export const useSetMapHoverInteraction = () => {
 export const useMapClick = (clickCallback: InteractionEventCallback) => {
   // const { updateFeatureState, cleanFeatureState } = useFeatureState(map)
   const onMapClick = useCallback(
-    (event: DeckLayerInteraction) => {
+    (event: InteractionEvent) => {
       if (!clickCallback) return
 
       const interactionEvent: InteractionEvent = {
