+++ libs/deck-layer-composer/src/interactions/types.ts
@@ -3,28 +3,27 @@ import {
   ContextPickingObject,
   FourwingsPickingObject,
 } from '@globalfishingwatch/deck-layers'
+import { DeckLayerInteractionFeature } from '../types'
 
 export type InteractionEventCallback = (event: InteractionEvent | null) => void
 
-export type ExtendedFeature = FourwingsPickingObject &
-  ContextPickingObject & {
-    layerId: string
-    generatorContextLayer?: ContextLayerId | null
-    datasetId?: string
-    promoteId?: string
-    id: string
-    value: any
-    geometry?: any
-    stopPropagation?: boolean
-    uniqueFeatureInteraction?: boolean
-    unit?: string
-    // TODO:deck review if this is needed anywhere else
-    // tile: {
-    //   x: number
-    //   y: number
-    //   z: number
-    // }
-  }
+export type ExtendedFeature = DeckLayerInteractionFeature & {
+  layerId: string
+  datasetId?: string
+  promoteId?: string
+  id: string
+  value: any
+  geometry?: any
+  stopPropagation?: boolean
+  uniqueFeatureInteraction?: boolean
+  unit?: string
+  // TODO:deck review if this is needed anywhere else
+  // tile: {
+  //   x: number
+  //   y: number
+  //   z: number
+  // }
+}
 
 export type InteractionEvent = {
   type: 'click' | 'hover'
