+++ libs/deck-layers/src/layers/context/context.types.ts
@@ -50,6 +50,10 @@ export type ContextFeature = Feature<Polygon | MultiPolygon, Record<string, any>
 // TODO:deck create this type in the proper deck class layer
 export type UserContextFeature = Feature<Geometry, Record<string, any>> & ContextFeatureProperties
 
-export type ContextPickingObject = ContextFeature | UserContextFeature
+export type ContextPickingObject = ContextFeature
+export type UserContextPickingObject = UserContextFeature
 
-export type ContextPickingInfo = PickingInfo<ContextPickingObject, { tile?: Tile2DHeader }>
+export type ContextPickingInfo = PickingInfo<
+  ContextPickingObject | UserContextPickingObject,
+  { tile?: Tile2DHeader }
+>
